+++ twitchio/models.py
@@ -85,9 +85,9 @@ class BitsLeaderboard:
 
     Attributes
     ------------
-    started_at: :class:`datetime.datetime`
+    started_at: Optional[:class:`datetime.datetime`]
         The time the leaderboard started.
-    ended_at: :class`datetime.datetime`
+    ended_at: Optional[:class:`datetime.datetime`]
         The time the leaderboard ended.
     leaders: List[:class:`BitLeaderboardUser`]
         The current leaders of the Leaderboard.
@@ -97,8 +97,10 @@ class BitsLeaderboard:
 
     def __init__(self, http: "TwitchHTTP", data: dict):
         self._http = http
-        self.started_at = datetime.datetime.fromisoformat(data["date_range"]["started_at"])
-        self.ended_at = datetime.datetime.fromisoformat(data["date_range"]["ended_at"])
+        self.started_at = (
+            parse_timestamp(data["date_range"]["started_at"]) if data["date_range"]["started_at"] else None
+        )
+        self.ended_at = parse_timestamp(data["date_range"]["ended_at"]) if data["date_range"]["ended_at"] else None
         self.leaders = [BitLeaderboardUser(http, x) for x in data["data"]]
 
     def __repr__(self):
@@ -333,7 +335,7 @@ class HypeTrainContribution:
         If type is ``SUBS``, aggregate total where 500, 1000, or 2500 represent tier 1, 2, or 3 subscriptions respectively.
         For example, if top contributor has gifted a tier 1, 2, and 3 subscription, total would be 4000.
     type: :class:`str`
-        Identifies the contribution method, either BITS or SUBS.
+        Identifies the contribution method, either BITS, SUBS or OTHER.
     user: :class:`~twitchio.PartialUser`
         The user making the contribution.
     """