+++ docs/changelog.rst
@@ -1,6 +1,20 @@
 :orphan:
 
 
+Master
+======
+- TwitchIO
+    - Additions
+        - Added optional ``started_at`` and ``ended_at`` arguments to :func:`~twitchio.PartialUser.fetch_clips`
+        - Updated docstring regarding new  HypeTrain contribution  method ``OTHER`` for :attr:`~twitchio.HypeTrainContribution.type`
+
+    - Bug fixes
+        - Fix :func:`~twitchio.PartialUser.fetch_bits_leaderboard` not handling ``started_at`` and :class:`~twitchio.BitsLeaderboard` not correctly parsing
+
+- ext.eventsub
+    - Additions
+        - Updated docs regarding new HypeTrain contribution method ``other`` for :attr:`~twitchio.ext.eventsub.HypeTrainContributor.type`
+
 2.5.0
 ======
 - TwitchIO